Fixes:    file size should be stored as qint64, not as an int, just like in qfileinfo.size()        RevBy:    Jesper
[git-p4: depot-paths = "//depot/qt/main/": change = 249405]

+++ src/gui/dialogs/qfileinfogatherer_p.h
@@ -35,7 +35,7 @@ public:
     QExtendedInformation() : size(0), fileType(System), isHidden(false),
                              isSymLink(false), caseSensitive(true) {}
 
-    int size;
+    qint64 size;
     QString displayType;
     QIcon icon;
     QDateTime lastModified;
